MODIFY : modulize routes/config.js
