+++ .gitignore
@@ -1,6 +1,5 @@
 
 db_config.js
-
 node_modules
 logs
